backend/python-model/requirements.txt
fastapi
uvicorn
python-chess==1.999
stockfish
pydantic
requests
